set request auth from optional endpoint url username/password
